ml/requirements.txt
pandas>=2.0
pyarrow>=14.0
tqdm>=4.65
sentence-transformers>=2.7
faiss-cpu>=1.7
torch>=2.1
openai>=1.0
